Refactor out the createLabelPipeline between views
