+++ src/VTKViewport/View3D.js
@@ -2,79 +2,13 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import vtkGenericRenderWindow from 'vtk.js/Sources/Rendering/Misc/GenericRenderWindow';
 import vtkWidgetManager from 'vtk.js/Sources/Widgets/Core/WidgetManager';
-import vtkImageData from 'vtk.js/Sources/Common/DataModel/ImageData';
-import vtkDataArray from 'vtk.js/Sources/Common/Core/DataArray';
-import vtkVolume from 'vtk.js/Sources/Rendering/Core/Volume';
-import vtkVolumeMapper from 'vtk.js/Sources/Rendering/Core/VolumeMapper';
 import vtkPaintFilter from 'vtk.js/Sources/Filters/General/PaintFilter';
 import vtkPaintWidget from 'vtk.js/Sources/Widgets/Widgets3D/PaintWidget';
-import vtkColorTransferFunction from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction';
-import vtkPiecewiseFunction from 'vtk.js/Sources/Common/DataModel/PiecewiseFunction';
 
 import ViewportOverlay from '../ViewportOverlay/ViewportOverlay.js';
 import { ViewTypes } from 'vtk.js/Sources/Widgets/Core/WidgetManager/Constants';
 import { createSub } from '../lib/createSub.js';
-
-// TODO: mostly duplicated in view2d...
-function createLabelPipeline(
-  backgroundImageData,
-  paintFilterLabelMapImageData
-) {
-  let labelMapData;
-
-  if (paintFilterLabelMapImageData) {
-    labelMapData = paintFilterLabelMapImageData;
-  } else {
-    // Create a labelmap image the same dimensions as our background volume.
-    labelMapData = vtkImageData.newInstance(
-      backgroundImageData.get('spacing', 'origin', 'direction')
-    );
-    labelMapData.setDimensions(backgroundImageData.getDimensions());
-    labelMapData.computeTransforms();
-
-    const values = new Uint8Array(backgroundImageData.getNumberOfPoints());
-    const dataArray = vtkDataArray.newInstance({
-      numberOfComponents: 1, // labelmap with single component
-      values,
-    });
-    labelMapData.getPointData().setScalars(dataArray);
-  }
-
-  const sampleDistance =
-    0.7 *
-    Math.sqrt(
-      labelMapData
-        .getSpacing()
-        .map(v => v * v)
-        .reduce((a, b) => a + b, 0)
-    );
-
-  const labelMap = {
-    actor: vtkVolume.newInstance(),
-    mapper: vtkVolumeMapper.newInstance(),
-    cfun: vtkColorTransferFunction.newInstance(),
-    ofun: vtkPiecewiseFunction.newInstance(),
-  };
-
-  labelMap.mapper.setSampleDistance(sampleDistance);
-
-  // labelmap pipeline
-  labelMap.actor.setMapper(labelMap.mapper);
-
-  // set up labelMap color and opacity mapping
-  labelMap.cfun.addRGBPoint(1, 0, 0, 1); // label "1" will be blue
-  labelMap.cfun.addRGBPoint(2, 1, 0, 0); // label "2" will be red
-  labelMap.cfun.addRGBPoint(3, 0, 1, 0); // label "3" will be green
-
-  labelMap.ofun.addPoint(0, 0);
-  labelMap.ofun.addPoint(1, 0.5);
-
-  labelMap.actor.getProperty().setRGBTransferFunction(0, labelMap.cfun);
-  labelMap.actor.getProperty().setScalarOpacity(0, labelMap.ofun);
-  labelMap.actor.getProperty().setInterpolationTypeToNearest();
-
-  return labelMap;
-}
+import createLabelPipeline from './createLabelPipeline';
 
 export default class View3D extends Component {
   static propTypes = {
@@ -89,12 +23,12 @@ export default class View3D extends Component {
     sliceNormal: PropTypes.array.isRequired,
     dataDetails: PropTypes.object,
     onCreated: PropTypes.func,
-    onDestroyed: PropTypes.func,
+    onDestroyed: PropTypes.func
   };
 
   static defaultProps = {
     painting: false,
-    sliceNormal: [0, 0, 1],
+    sliceNormal: [0, 0, 1]
   };
 
   constructor(props) {
@@ -109,13 +43,13 @@ export default class View3D extends Component {
       labelmap: createSub(),
       paint: createSub(),
       paintStart: createSub(),
-      paintEnd: createSub(),
+      paintEnd: createSub()
     };
   }
 
   componentDidMount() {
     this.genericRenderWindow = vtkGenericRenderWindow.newInstance({
-      background: [0, 0, 0],
+      background: [0, 0, 0]
     });
 
     this.genericRenderWindow.setContainer(this.container.current);
@@ -182,7 +116,7 @@ export default class View3D extends Component {
         filters,
         actors,
         volumes,
-        _component: this,
+        _component: this
       };
 
       this.props.onCreated(api);
@@ -254,7 +188,8 @@ export default class View3D extends Component {
       const labelmapImageData = this.props.paintFilterLabelMapImageData;
       const labelmap = createLabelPipeline(
         this.props.paintFilterBackgroundImageData,
-        labelmapImageData
+        labelmapImageData,
+        true
       );
 
       this.labelmap = labelmap;
@@ -350,7 +285,7 @@ export default class View3D extends Component {
 
     let voi = {
       windowCenter: 0,
-      windowWidth: 0,
+      windowWidth: 0
     };
 
     if (this.pipeline) {
@@ -364,7 +299,7 @@ export default class View3D extends Component {
 
       voi = {
         windowCenter,
-        windowWidth,
+        windowWidth
       };
     }
