+++ src/VTKViewport/createLabelPipeline.js
@@ -1,0 +1,67 @@
+import vtkImageData from 'vtk.js/Sources/Common/DataModel/ImageData';
+import vtkDataArray from 'vtk.js/Sources/Common/Core/DataArray';
+import vtkVolume from 'vtk.js/Sources/Rendering/Core/Volume';
+import vtkVolumeMapper from 'vtk.js/Sources/Rendering/Core/VolumeMapper';
+import vtkColorTransferFunction from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction';
+import vtkPiecewiseFunction from 'vtk.js/Sources/Common/DataModel/PiecewiseFunction';
+
+export default function createLabelPipeline(
+  backgroundImageData,
+  paintFilterLabelMapImageData,
+  useSampleDistance = false
+) {
+  let labelMapData;
+
+  if (paintFilterLabelMapImageData) {
+    labelMapData = paintFilterLabelMapImageData;
+  } else {
+    // Create a labelmap image the same dimensions as our background volume.
+    labelMapData = vtkImageData.newInstance(
+      backgroundImageData.get('spacing', 'origin', 'direction')
+    );
+    labelMapData.setDimensions(backgroundImageData.getDimensions());
+    labelMapData.computeTransforms();
+
+    const values = new Uint8Array(backgroundImageData.getNumberOfPoints());
+    const dataArray = vtkDataArray.newInstance({
+      numberOfComponents: 1, // labelmap with single component
+      values
+    });
+    labelMapData.getPointData().setScalars(dataArray);
+  }
+  if (useSampleDistance) {
+    const sampleDistance =
+      0.7 *
+      Math.sqrt(
+        labelMapData
+          .getSpacing()
+          .map(v => v * v)
+          .reduce((a, b) => a + b, 0)
+      );
+
+    labelMap.mapper.setSampleDistance(sampleDistance);
+  }
+
+  const labelMap = {
+    actor: vtkVolume.newInstance(),
+    mapper: vtkVolumeMapper.newInstance(),
+    cfun: vtkColorTransferFunction.newInstance(),
+    ofun: vtkPiecewiseFunction.newInstance()
+  };
+
+  // labelmap pipeline
+  labelMap.actor.setMapper(labelMap.mapper);
+
+  // set up labelMap color and opacity mapping
+  labelMap.cfun.addRGBPoint(1, 0, 0, 1); // label '1' will be blue
+  labelMap.cfun.addRGBPoint(2, 1, 0, 0); // label '2' will be red
+  labelMap.cfun.addRGBPoint(3, 0, 1, 0); // label '3' will be green
+  labelMap.ofun.addPoint(0, 0);
+  labelMap.ofun.addPoint(1, 0.5);
+
+  labelMap.actor.getProperty().setRGBTransferFunction(0, labelMap.cfun);
+  labelMap.actor.getProperty().setScalarOpacity(0, labelMap.ofun);
+  labelMap.actor.getProperty().setInterpolationTypeToNearest();
+
+  return labelMap;
+}
